drm/msm: Reduce latency while completing non-blocking commits
In order to complete a commit, we must first wait for the previous
commit to finish up. This is done by sleeping, during which time the CPU
can enter a deep idle state and take a while to finish processing the
commit after the wait is over. We can alleviate this by optimistically
assuming that the kthread this commit worker is running on won't migrate
to another CPU partway through. We only do this for the non-blocking
case where the commit completion is done in an asynchronous worker
because the generic DRM code already does this for atomic ioctls. Since
non-blocking commits are the common case, preemptively kick the prime
CPU out of idle when the atomic ioctl runs so that it'll be ready to
process the DRM IRQ and display commit kthreads.

Change-Id: I2a6a0a862ef340b444c58cf1a5fb9b00b0985bd2
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: Adithya R <gh0strider.2k18.reborn@gmail.com>
Signed-off-by: rk134 <me@rk134.cf>